+++ R/netcdf-utils.R
@@ -14,8 +14,10 @@ nc_get.character <- function(x, v, test = FALSE, ...) {
     if (!is.null(val$result)) return(val$result)
   } else {
   con4 <- NULL
+  ## issue #119
+  suppressWarnings(
   con4 <- ncdf4::nc_open(x, readunlim = FALSE, verbose = FALSE, 
-                         auto_GMT = FALSE, suppress_dimvals = TRUE)
+                         auto_GMT = FALSE, suppress_dimvals = TRUE))
   on.exit(ncdf4::nc_close(con4), add = TRUE)
   safe_get4 <- purrr::safely(nc_get.ncdf4)
   val <- safe_get4(con4, v)